make the changes given by the reviewr
