+++ .rubocop.yml
@@ -52,3 +52,5 @@ Style/HashTransformKeys:
   Enabled: false
 Style/HashTransformValues:
   Enabled: false
+  Layout/EndOfLine:
+   Enabled: False
