--- School/rental.rb
@@ -1,22 +1,0 @@
-class Rental
-  attr_accessor :date
-  attr_reader :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @person = person
-    person.rental << self
-    @book = book
-    book.rental << self
-  end
-
-  def book=(book)
-    @book = book
-    book.rentals.push(self) unless book.rentals.include?(self)
-  end
-
-  def person=(person)
-    @person = person
-    person.rentals.push(self) unless person.rentals.include?(self)
-  end
-end
