+++ book.rb
@@ -2,7 +2,7 @@ class Book
   attr_accessor :title, :author
   attr_reader :rental
 
-  def initalize(title, author)
+  def initialize(title, author)
     @title = title
     @author = author
     @rental = []
